refactor: decompose 4 oversized tools modules into siblings (#2077)
