+++ data/feature_index.json
@@ -1199,6 +1199,6 @@
       ]
     }
   ],
-  "generated_at": "2026-06-08T14:57:13.957188Z",
+  "generated_at": "2026-06-08T18:35:35.143924Z",
   "schema_version": 1
 }
